requirements.txt
environs

# Apache Airflow
apache-airflow

# Библиотека для работы с переменными окружения
python-dotenv

# Библиотеки для работы с PostgreSQL
psycopg2-binary

# Дополнительные библиотеки для работы с данными
pandas
numpy

# Библиотека для работы с Kaggle (если потребуется в будущем)
kaggle